Do not ignore k* and c* parameters if mass==0
Visco-Elastic model ignored strength and damping parameter,
if the mass body is zero. It happens during facet-sphere
interaction, so the facets mat is ignored (k* and c* multiplied
by 0).

This commit sets temporarly the mass of zero-mass body equal
to the mass of another body.

Thanks to Medack for pointing that out and discussion.